// web/src/generated/models/evaluation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Aidash API
 * Management API for the Aidash agent mesh.
 * OpenAPI spec version: 0.1.0
 */
import type { Resource } from "./resource.ts";

export interface Evaluation {
  action: string;
  environment?: unknown;
  resource: Resource;
  subject: string;
}