// web/src/generated/models/policyBundle.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Aidash API
 * Management API for the Aidash agent mesh.
 * OpenAPI spec version: 0.1.0
 */
import type { Policy } from "./policy.ts";
import type { PolicyBundleGroups } from "./policyBundleGroups.ts";
import type { PolicyBundleRoles } from "./policyBundleRoles.ts";
import type { PolicyBundleSubjects } from "./policyBundleSubjects.ts";

export interface PolicyBundle {
  groups?: PolicyBundleGroups;
  policies?: Policy[];
  roles?: PolicyBundleRoles;
  subjects?: PolicyBundleSubjects;
  tenant: string;
}